Move external API/console impls to subpackages and exclude from JaCoCo
