+++ src/main/kotlin/werewolf/ai/anthropic/AnthropicLanguageModel.kt
@@ -1,10 +1,12 @@
-package werewolf.ai
+package werewolf.ai.anthropic
 
 import com.anthropic.client.AnthropicClient
 import com.anthropic.client.okhttp.AnthropicOkHttpClient
 import com.anthropic.models.messages.CacheControlEphemeral
 import com.anthropic.models.messages.MessageCreateParams
 import com.anthropic.models.messages.TextBlockParam
+import werewolf.ai.Completion
+import werewolf.ai.LanguageModel
 
 class AnthropicLanguageModel(
     private val model: String = DEFAULT_MODEL,
